backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
sqlalchemy==2.0.35
google-generativeai==0.8.3
httpx==0.27.2
pyautogui==0.9.54
pyperclip==1.9.0
psutil==6.1.0
pillow==11.0.0
playwright==1.48.0
edge-tts
pygame
pyttsx3==2.90
SpeechRecognition==3.10.4
pyaudio==0.2.14
pydantic-settings==2.6.1
aiofiles==24.1.0
python-multipart==0.0.12
